Properties fixup
Mostly this just changes the customized properties to use the _item_prop
function.

I don't think it is the libraries responsibility to try and prevent
people from doing things like deleting properties or putting custom
handling on properties. If we want to build helpers we should build them
on top of the RFC primitives and give users access to both.

This also adds some documentation which was copied from the RFC.